refactor(apps): placement tests name the invariant they prove; the sandbox fake drops its off-seam simulator
Four test-estate cleanups in @vendoai/apps. No behaviour change anywhere:
renames, comment rewrites, consolidation, deletions.

- `placement-risk.test.ts` / `placement-risk-2.test.ts` were review-session
  transcripts, not suites: a filename that is a round number, a header
  announcing "ADVERSARIAL round 2", cases citing a "Checker round-1 ruling",
  and every describe/it worded as a live defect while the assertion proved
  the opposite. The suite printed "keeps naming an app the caller lost
  access to" over `toEqual([])`. Their comments had also gone false — one
  said the read path "does not trim at all" next to `input.slots?.map(
  requireSlot)`, another titled the eviction receipt "read-then-write, with
  no CAS" next to a CAS loop. All nine cases are folded into
  `placement-runtime.test.ts`, renamed after the behaviour they prove and
  re-commented with the invariant rather than the bug that once existed.
  Five placement-named test files become three.

- `FakeSandboxMachine` carried an exec simulator (a regex parsing `printf`
  shell strings, another parsing `fetch(...).then(() => process.exit(N))`)
  and a hardcoded one-pixel PNG screenshotter. Neither `exec` nor
  `screenshot` is on the `SandboxMachine` seam — the seam's own header says
  exec left it — and nothing called either but the fake's own test of its
  own simulation. The fake itself stays: it is a full SandboxAdapter used
  by six real test files.

- `engine.live.test.ts` built `createApps` with no `screen` assembler and
  then called `runtime.create(...)`, which throws on exactly that path. It
  could not pass whatever the live model returned, and any developer with
  ANTHROPIC_API_KEY exported got a hard red on `pnpm test:affected`.
