+++ packages/apps/src/testing/fake-sandbox.test.ts
@@ -130,21 +130,4 @@ describe("fakeSandbox v2 seam semantics", () => {
   it("adapter.destroy rejects a ref from another provider", async () => {
     await expect(fakeSandbox().destroy("e2b:v2:whatever")).rejects.toThrow(/fake sandbox snapshot ref/);
   });
-
-  it("simulates the egress allowlist in exec fetches", async () => {
-    const adapter = fakeSandbox();
-    const machine = await adapter.create({
-      env: { PORT: "8080" },
-      allowedDomains: ["api.example.com", "*.wild.example"],
-    });
-    expect(machine.allowedDomains).toEqual(["api.example.com", "*.wild.example"]);
-    const allowed = await machine.exec(
-      "timeout 5 node -e \"fetch('https://api.example.com').then(() => process.exit(0)).catch(() => process.exit(1))\"",
-    );
-    expect(allowed.code).toBe(0);
-    const blocked = await machine.exec(
-      "timeout 5 node -e \"fetch('https://evil.example').then(() => process.exit(0)).catch(() => process.exit(1))\"",
-    );
-    expect(blocked.code).toBe(1);
-  });
 });
